+++ fabric/src/main/java/com/chefmooon/frightsdelight/data/fabric/ItemTagGenerator.java
@@ -59,7 +59,14 @@ public class ItemTagGenerator extends FabricTagProvider.ItemTagProvider {
 //                .add(FrightsDelightItemsImpl.COOKIE_COBWEB)
 //                .add(FrightsDelightItemsImpl.COOKIE_GHAST_TEAR)
 //                .add(FrightsDelightItemsImpl.COOKIE_SOUL_BERRY)
-//                .add(FrightsDelightItemsImpl.COOKIE_WITHER_BERRY);
+//                .add(FrightsDelightItemsImpl.COOKIE_WITHER_BERRY)
+//                .add(FrightsDelightItemsImpl.ROTTEN_FLESH_PIE)
+//                .add(FrightsDelightItemsImpl.SLIMEAPPLE_PIE)
+//                .add(FrightsDelightItemsImpl.SPIDEREYE_PIE)
+//                .add(FrightsDelightItemsImpl.GHASTTEAR_PIE)
+//                .add(FrightsDelightItemsImpl.SOUL_BERRY_CHEESECAKE)
+//                .add(FrightsDelightItemsImpl.WITHER_BERRY_CHEESECAKE);
+//                .add(FrightsDelightItemsImpl.COBWEB_PIE)
     }
 
     private void registerCompatibilityItemTags() {
@@ -86,7 +93,14 @@ public class ItemTagGenerator extends FabricTagProvider.ItemTagProvider {
                 .add(FrightsDelightItemsImpl.SOUP_COBWEB)
                 .add(FrightsDelightItemsImpl.SOUP_GHAST_TEAR)
                 .add(FrightsDelightItemsImpl.SOUP_SOUL_BERRY)
-                .add(FrightsDelightItemsImpl.SOUP_WITHER_BERRY);
+                .add(FrightsDelightItemsImpl.SOUP_WITHER_BERRY)
+                .add(FrightsDelightItemsImpl.ROTTEN_FLESH_PIE)
+                .add(FrightsDelightItemsImpl.SLIMEAPPLE_PIE)
+                .add(FrightsDelightItemsImpl.SPIDEREYE_PIE)
+                .add(FrightsDelightItemsImpl.GHASTTEAR_PIE)
+                .add(FrightsDelightItemsImpl.SOUL_BERRY_CHEESECAKE)
+                .add(FrightsDelightItemsImpl.WITHER_BERRY_CHEESECAKE)
+                .add(FrightsDelightItemsImpl.COBWEB_PIE);
 
         // Create: Craft's and Additions
         getOrCreateTagBuilder(CompatibilityTags.CREATE_ADDITION_PLANT_FOODS)
@@ -143,6 +157,9 @@ public class ItemTagGenerator extends FabricTagProvider.ItemTagProvider {
                 .add(FrightsDelightItemsImpl.PUNCH_ROTTEN_FLESH)
                 .add(FrightsDelightItemsImpl.PUNCH_SPIDEREYE)
                 .add(FrightsDelightItemsImpl.PUNCH_SLIMEAPPLE)
+                .add(FrightsDelightItemsImpl.ROTTEN_FLESH_PIE_SLICE)
+                .add(FrightsDelightItemsImpl.SLIMEAPPLE_PIE_SLICE)
+                .add(FrightsDelightItemsImpl.SPIDEREYE_PIE_SLICE)
         ;
 
         getOrCreateTagBuilder(CompatibilityTags.ORIGINS_IGNORE_DIET)
@@ -154,6 +171,8 @@ public class ItemTagGenerator extends FabricTagProvider.ItemTagProvider {
                 .add(FrightsDelightItemsImpl.SOUP_COBWEB)
                 .add(FrightsDelightItemsImpl.PUNCH_GHASTTEAR)
                 .add(FrightsDelightItemsImpl.PUNCH_COBWEB)
+                .add(FrightsDelightItemsImpl.GHASTTEAR_PIE_SLICE)
+                .add(FrightsDelightItemsImpl.COBWEB_PIE_SLICE)
         ;
     }
 }