+++ fabric/src/main/java/com/chefmooon/frightsdelight/data/fabric/RecipeGenerator.java
@@ -1,9 +1,12 @@
 package com.chefmooon.frightsdelight.data.fabric;
 
 import com.chefmooon.frightsdelight.common.registry.fabric.FrightsDelightItemsImpl;
+import com.chefmooon.frightsdelight.common.tag.CommonTags;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricRecipeProvider;
+import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.data.recipes.*;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.Items;
 import vectorwing.farmersdelight.common.registry.ModItems;
@@ -43,6 +46,14 @@ public class RecipeGenerator extends FabricRecipeProvider {
         punchbowlFromPunch(FrightsDelightItemsImpl.PUNCHBOWL_SOUL_BERRY, FrightsDelightItemsImpl.PUNCH_SOUL_BERRY, exporter);
         punchbowlFromPunch(FrightsDelightItemsImpl.PUNCHBOWL_WITHER_BERRY, FrightsDelightItemsImpl.PUNCH_WITHER_BERRY, exporter);
         punchbowlFromPunch(FrightsDelightItemsImpl.PUNCHBOWL_COBWEB, FrightsDelightItemsImpl.PUNCH_COBWEB, exporter);
+
+        pieRecipe(FrightsDelightItemsImpl.ROTTEN_FLESH_PIE, FrightsDelightItemsImpl.ROTTEN_FLESH_PIE_SLICE, Items.ROTTEN_FLESH, exporter);
+        pieRecipe(FrightsDelightItemsImpl.SLIMEAPPLE_PIE, FrightsDelightItemsImpl.SLIMEAPPLE_PIE_SLICE, FrightsDelightItemsImpl.SLIMEAPPLE_PIE, exporter);
+        pieRecipe(FrightsDelightItemsImpl.SPIDEREYE_PIE, FrightsDelightItemsImpl.SPIDEREYE_PIE_SLICE, Items.SPIDER_EYE, exporter);
+        pieRecipe(FrightsDelightItemsImpl.GHASTTEAR_PIE, FrightsDelightItemsImpl.GHASTTEAR_PIE_SLICE, Items.GHAST_TEAR, exporter);
+        pieRecipe(FrightsDelightItemsImpl.SOUL_BERRY_CHEESECAKE, FrightsDelightItemsImpl.SOUL_BERRY_CHEESECAKE_SLICE, FrightsDelightItemsImpl.SOUL_BERRY, exporter);
+        pieRecipe(FrightsDelightItemsImpl.WITHER_BERRY_CHEESECAKE, FrightsDelightItemsImpl.WITHER_BERRY_CHEESECAKE_SLICE, FrightsDelightItemsImpl.WITHER_BERRY, exporter);
+        pieRecipe(FrightsDelightItemsImpl.COBWEB_PIE, FrightsDelightItemsImpl.COBWEB_PIE_SLICE, Items.COBWEB, exporter);
     }
 
     private static void crateToIngredient(Item crate, Item ingredient, Consumer<FinishedRecipe> exporter) {
@@ -70,5 +81,29 @@ public class RecipeGenerator extends FabricRecipeProvider {
                 .unlockedBy(RecipeProvider.getHasName(punch), RecipeProvider.has(punch))
                 .showNotification(false)
                 .save(exporter, RecipeProvider.getSimpleRecipeName(punchbowl));
+    }
+
+    private static void pieRecipe(Item pie, Item pieSlice, Item mainIngredient, Consumer<FinishedRecipe> exporter) {
+        Item crust = BuiltInRegistries.ITEM.get(new ResourceLocation("farmersdelight", "pie_crust"));
+        // This must be made manually due to milk tag difference fabric -> c:milks | forge -> forge:milk
+//        ShapedRecipeBuilder.shaped(RecipeCategory.MISC, pie)
+//                .pattern("AAA")
+//                .pattern("AAA")
+//                .pattern("BCB")
+//                .define('A', mainIngredient)
+//                .define('B', CommonTags.C_FOODS_MILK)
+//                .define('C', crust)
+//                .unlockedBy("has_pie_crust", RecipeProvider.has(crust))
+//                .group("frd_" + pie.getDescriptionId().replace("block.frightsdelight.", ""))
+//                .save(exporter, RecipeProvider.getSimpleRecipeName(pie));
+
+        ShapedRecipeBuilder.shaped(RecipeCategory.FOOD, pie)
+                .pattern("AA")
+                .pattern("AA")
+                .define('A', pieSlice)
+                .unlockedBy(RecipeProvider.getHasName(pieSlice), RecipeProvider.has(pieSlice))
+                .group("frd_" + pie.getDescriptionId().replace("block.frightsdelight.", ""))
+                .save(exporter, RecipeProvider.getSimpleRecipeName(pie) + "_from_slices");
+
     }
 }