+++ changelog.md
@@ -7,6 +7,22 @@ NeoForge, and Fabric
 
 ***
 
+### Added
+
+- New Food
+  - Rotten Flesh Pie (+Slice)
+  - Slime Apple Pie (+Slice)
+  - Spider Eye Pie (+Slice)
+  - Ghast Tear Pie (+Slice)
+  - Soul Berry Cheesecake (+Slice)
+  - Wither Berry Cheesecake (+Slice)
+  - Cobweb Pie (+Slice)
+- Origins Tags
+    - Origins Meat: Slice of Rotten Flesh Pie, Slice of Slime-Apple Pie, Slice of Spider Eye Pie
+    - Ignore Diet: Ghast Tear Pie Slice, Cobweb Pie Slice
+- Create Tag
+  - Upright on Belt: Rotten Flesh Pie, Slime-Apple Pie, Spider Eye Pie, Ghast Tear Pie, Soul Berry Cheesecake, Wither Berry Cheesecake, Cobweb Pie
+
 ### Changed
 
 - [1.20.1+] Bone Shards now require a short charge time before they can be thrown.